requirements.txt
Flask
gunicorn
Pillow
google-cloud-vision